Add initial support for dependency injection
