+++ composer.json
@@ -10,7 +10,8 @@
         }
     ],
     "require": {
-        "php": ">=7.1"
+        "php": ">=7.1",
+        "league/container": "^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^7.0"
@@ -19,12 +20,11 @@
         "files": [
             "./src/Language/blockStringValue.php",
             "./src/Language/helpers.php",
-            "./src/Language/parse.php",
-            "./src/Language/print.php",
             "./src/Type/definition.php",
             "./src/Type/directives.php",
             "./src/Type/scalars.php",
-            "./src/Util/helpers.php"
+            "./src/Util/helpers.php",
+            "./src/facades.php"
         ],
         "psr-4": {
             "Digia\\GraphQL\\": "./src"
@@ -34,5 +34,6 @@
         "psr-4": {
             "Digia\\GraphQL\\Test\\": "./tests"
         }
-    }
+    },
+    "minimum-stability": "dev"
 }
