--- src/GraphQL.php
@@ -1,14 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: chris
- * Date: 02/03/2018
- * Time: 8.28
- */
-
-namespace Digia\GraphQL;
-
-class GraphQL
-{
-
-}
