+++ src/Type/directives.php
@@ -1,12 +1,8 @@
 <?php
 
-use Digia\GraphQL\Language\AST\DirectiveLocationEnum;
 use Digia\GraphQL\Type\Definition\Directive;
 use Digia\GraphQL\Type\Definition\DirectiveInterface;
-use function Digia\GraphQL\Type\GraphQLBoolean;
-use function Digia\GraphQL\Type\GraphQLDirective;
-use function Digia\GraphQL\Type\GraphQLNonNull;
-use function Digia\GraphQL\Type\GraphQLString;
+use function Digia\GraphQL\graphql;
 use function Digia\GraphQL\Util\arraySome;
 
 /**
@@ -15,29 +11,7 @@ use function Digia\GraphQL\Util\arraySome;
  */
 function GraphQLIncludeDirective(): Directive
 {
-    static $instance = null;
-
-    if ($instance === null) {
-        $instance = GraphQLDirective([
-            'name'        => 'include',
-            'description' =>
-                'Directs the executor to include this field or fragment only when ' .
-                'the `if` argument is true.',
-            'locations'   => [
-                DirectiveLocationEnum::FIELD,
-                DirectiveLocationEnum::FRAGMENT_SPREAD,
-                DirectiveLocationEnum::INLINE_FRAGMENT,
-            ],
-            'args'        => [
-                'if ' => [
-                    'type'        => GraphQLNonNull(GraphQLBoolean()),
-                    'description' => 'Included when true.',
-                ],
-            ],
-        ]);
-    }
-
-    return $instance;
+    return graphql()->get('GraphQLIncludeDirective');
 }
 
 /**
@@ -46,29 +20,7 @@ function GraphQLIncludeDirective(): Directive
  */
 function GraphQLSkipDirective(): Directive
 {
-    static $instance = null;
-
-    if ($instance === null) {
-        $instance = GraphQLDirective([
-            'name'        => 'skip',
-            'description' =>
-                'Directs the executor to skip this field or fragment when the `if` ' .
-                'argument is true.',
-            'locations'   => [
-                DirectiveLocationEnum::FIELD,
-                DirectiveLocationEnum::FRAGMENT_SPREAD,
-                DirectiveLocationEnum::INLINE_FRAGMENT,
-            ],
-            'args'        => [
-                'if' => [
-                    'type'        => GraphQLNonNull(GraphQLBoolean()),
-                    'description' => 'Skipped when true.',
-                ],
-            ],
-        ]);
-    }
-
-    return $instance;
+    return graphql()->get('GraphQLSkipDirective');
 }
 
 const DEFAULT_DEPRECATION_REASON = 'No longer supported';
@@ -79,30 +31,7 @@ const DEFAULT_DEPRECATION_REASON = 'No longer supported';
  */
 function GraphQLDeprecatedDirective(): Directive
 {
-    static $instance = null;
-
-    if ($instance === null) {
-        $instance = GraphQLDirective([
-            'name'        => 'deprecated',
-            'description' => 'Marks an element of a GraphQL schema as no longer supported.',
-            'locations'   => [
-                DirectiveLocationEnum::FIELD_DEFINITION,
-                DirectiveLocationEnum::ENUM_VALUE,
-            ],
-            'args'        => [
-                'reason' => [
-                    'type'         => GraphQLString(),
-                    'description'  =>
-                        'Explains why this element was deprecated, usually also including a ' .
-                        'suggestion for how to access supported similar data. Formatted ' .
-                        'in [Markdown](https://daringfireball.net/projects/markdown/).',
-                    'defaultValue' => DEFAULT_DEPRECATION_REASON,
-                ],
-            ]
-        ]);
-    }
-
-    return $instance;
+    return graphql()->get('GraphQLDeprecatedDirective');
 }
 
 /**
